+++ ipfrag/iprbexplicit.c
@@ -94,7 +94,7 @@ rb_explicit_reassctx_reassemble(struct allocif *loc, struct rb_explicit_reassctx
     len = ip_total_len(iporig) - ip_hdr_len(iporig);
     if (first + len - 1 > ctx->most_restricting_last)
     {
-      memcpy(pay2 + first, payorig, 1 + ctx->most_restricting_last - first);
+      memcpy(pay2 + first, payorig, 1 + (size_t)ctx->most_restricting_last - first);
     }
     else
     {
